etiqueta_generator/requirements.txt
pandas>=1.3.0
reportlab>=3.6.0
python-barcode>=0.14.0
pillow>=8.3.0
